% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbb.R
\docType{data}
\name{cbb8}
\alias{cbb8}
\alias{cbb12}
\title{Colorblind-friendly palletes}
\format{cbb8 is a character vector with 8 values, and cbb12 is a tibble with 4 columns and 12 rows.
\describe{
   \item{cbb8}{A vector with hex codes for 8 colors.}
   \item{cbb12$col}{Hex strings specifying the colors.}
   \item{cbb12$dark2light}{Ordering of col from dark to light.}
   \item{cbb12$contrast}{Ordering of col optimized for contrast.}
   \item{cbb12$tritanopia}{Identifies three colors that are indistinguishable for individuals with tritanopia.}
   \item{cbb12$dropTrit}{Identifies the three colors to drop when not including colors indistinguishible for individuals with tritanopia.}
}}
\source{
\url{http://www.cookbook-r.com/Graphs/Colors_(ggplot2)/}

\url{http://mkweb.bcgsc.ca/colorblind/img/colorblindness.palettes.simple.png}
}
\usage{
cbb8

cbb12
}
\description{
Colorblind-friendly palletes
}
\note{
cbb12 is best accessed using the \code{\link{CBB12}} function.
}
\keyword{datasets}
