% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghostwriter.R
\name{ghostwriter}
\alias{ghostwriter}
\title{Write an Apple script to type some text into RStudio}
\usage{
ghostwriter(file, lines = NULL, speed = 0.1,
  script = "ghostwrite.scpt", application = "RStudio")
}
\arguments{
\item{file}{A file with the desired text to write}

\item{lines}{A numeric vector specifying which lines to type. If unspecified, the entire file will be typed out.}

\item{speed}{A number parameterizing the speed at which the script will type.}

\item{script}{A path to the location where the script will be written.}

\item{application}{The application where the keyboard input should go.}
}
\value{
Returns the script invisibly.
}
\description{
This is used to write an Apple Script that will type some text into RStudio. It's primary purpose is for use in recording videos when I want to show live "typing" without any typos or delays while live typing.
}
\note{
This currently only allows hard returns as special characters. Any other characters will be entered exactly as they appear in the file.
}
